Some aesthetic changes...
* Use python native logging mod
* Support two levels of verbosity (default is more or less what it was)
* Clean up the structure of the model directory
