+++ .travis.yml
@@ -13,6 +13,10 @@ env:
   - DOCKER_HUB_USERNAME=otakup0pe
   - DOCKER_HUB_EMAIL=jonafree@gmail.com
   - secure: ep1otFA9phmSyP4R9tPbu/1tBEN72jIEWxG5xQam63namK4Fwy5s1VSB6hv37JN48E/MVRYzgQEDBuA6K5LLCtWJdIFYGAiniduVPBGDncPprp/4vAzXc6YcdBpPNI0/yLWorz3/xNsjZk7G9UrojJu2KdAHhJ+mFZhGesXuPyYav4hgowp88AjGQSHq5RQyK6Drht5OeuEV6tTs8lLEzqOYZWonWNuTCxE38LhfiTn4MWZijCPB9YIOfA0YNN9RX9vaoocXm5OOt1jhXDCaP/pVblC0A1yLiMT0qBmMuWy/Col/22UmND3+Cumva4Y+fY7heetNHEK2FJnTHfSG9Lsd149JqNfz0CfV/BvFNCp12TS7hAV2Mmkb6mf8Gi/zea9mA/3DwiIkjfcCK8Hkfmna4OmkFyrEGYergEIr1rGgo/zJxE1R+CZHwLDPvuY9yimQ51Y9d63NweHxt3b5F8+MGPAN9yWsGdFWcoo8xdo6M0otHurFx0RpFoOtBkljMK8OblE+ncIj1JpNW0zzDktlEHPeXpzN67HgcKw+FwfSedm5RDgYRNuu2e8BhYkJDwu04fYvxDIJklxdO+yCa7Ul4FL8fNuCxOS6tQAtAdaEye3rxcb8mUolMD3P44GLADhy4QC3Fe7F0BXUKwdKl4/wkXUN1o10oMREQZiE33Q=
+  matrix:
+    - VAULT_VERSION=0.6.2
+    - VAULT_VERSION=0.7.0
+    - VAULT_VERSION=0.7.4
 deploy:
   skip_cleanup: true  
   provider: pypi
