+++ aomi/__init__.py
@@ -1,6 +1,4 @@
 """The Tool Known as aomi"""
 # These are loaded here because metaprogramming
 # pylint: disable=unused-import
-import aomi.model.generic
-import aomi.model.auth
-import aomi.model.aws
+import aomi.model
